Abstracts the closure, moves instruction and command to process
