+++ synthesizer/src/process/command/contains.rs
