+++ synthesizer/src/process/command/remove.rs
