+++ synthesizer/src/process/instruction/bytes.rs
