+++ synthesizer/src/process/instruction/operand/bytes.rs
