+++ synthesizer/src/process/instruction/operation/commit.rs
@@ -344,9 +344,9 @@ impl<N: Network, const VARIANT: u8> ToBytes for CommitInstruction<N, VARIANT> {
 #[cfg(test)]
 mod tests {
     use super::*;
-    use crate::{
-        process::Stack,
-        program::test_helpers::{sample_finalize_registers, sample_registers},
+    use crate::process::{
+        instruction::test_helpers::{sample_finalize_registers, sample_registers},
+        Stack,
     };
     use circuit::{AleoV0, Eject};
     use console::{network::Testnet3, program::Identifier};
