+++ synthesizer/src/process/instruction/operation/is.rs
@@ -12,37 +12,49 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-use crate::{Opcode, Operand, RegistersLoad, RegistersLoadCircuit, StackMatches, StackProgram};
+use crate::{
+    Opcode,
+    Operand,
+    RegistersLoad,
+    RegistersLoadCircuit,
+    RegistersStore,
+    RegistersStoreCircuit,
+    StackMatches,
+    StackProgram,
+};
 use console::{
     network::prelude::*,
-    program::{Register, RegisterType},
+    program::{Literal, LiteralType, Plaintext, PlaintextType, Register, RegisterType, Value},
+    types::Boolean,
 };
 
-/// Asserts two operands are equal to each other.
-pub type AssertEq<N> = AssertInstruction<N, { Variant::AssertEq as u8 }>;
-/// Asserts two operands are **not** equal to each other.
-pub type AssertNeq<N> = AssertInstruction<N, { Variant::AssertNeq as u8 }>;
+/// Computes whether `first` equals `second` as a boolean, storing the outcome in `destination`.
+pub type IsEq<N> = IsInstruction<N, { Variant::IsEq as u8 }>;
+/// Computes whether `first` does **not** equals `second` as a boolean, storing the outcome in `destination`.
+pub type IsNeq<N> = IsInstruction<N, { Variant::IsNeq as u8 }>;
 
 enum Variant {
-    AssertEq,
-    AssertNeq,
+    IsEq,
+    IsNeq,
 }
 
-/// Asserts an operation on two operands.
+/// Computes an equality operation on two operands, and stores the outcome in `destination`.
 #[derive(Clone, PartialEq, Eq, Hash)]
-pub struct AssertInstruction<N: Network, const VARIANT: u8> {
+pub struct IsInstruction<N: Network, const VARIANT: u8> {
     /// The operands.
     operands: Vec<Operand<N>>,
+    /// The destination register.
+    destination: Register<N>,
 }
 
-impl<N: Network, const VARIANT: u8> AssertInstruction<N, VARIANT> {
+impl<N: Network, const VARIANT: u8> IsInstruction<N, VARIANT> {
     /// Returns the opcode.
     #[inline]
     pub const fn opcode() -> Opcode {
         match VARIANT {
-            0 => Opcode::Assert("assert.eq"),
-            1 => Opcode::Assert("assert.neq"),
-            _ => panic!("Invalid 'assert' instruction opcode"),
+            0 => Opcode::Is("is.eq"),
+            1 => Opcode::Is("is.neq"),
+            _ => panic!("Invalid 'is' instruction opcode"),
         }
     }
 
@@ -50,7 +62,7 @@ impl<N: Network, const VARIANT: u8> AssertInstruction<N, VARIANT> {
     #[inline]
     pub fn operands(&self) -> &[Operand<N>] {
         // Sanity check that the operands is exactly two inputs.
-        debug_assert!(self.operands.len() == 2, "Assert operations must have two operands");
+        debug_assert!(self.operands.len() == 2, "Instruction '{}' must have two operands", Self::opcode());
         // Return the operands.
         &self.operands
     }
@@ -58,17 +70,17 @@ impl<N: Network, const VARIANT: u8> AssertInstruction<N, VARIANT> {
     /// Returns the destination register.
     #[inline]
     pub fn destinations(&self) -> Vec<Register<N>> {
-        vec![]
+        vec![self.destination.clone()]
     }
 }
 
-impl<N: Network, const VARIANT: u8> AssertInstruction<N, VARIANT> {
+impl<N: Network, const VARIANT: u8> IsInstruction<N, VARIANT> {
     /// Evaluates the instruction.
     #[inline]
     pub fn evaluate(
         &self,
         stack: &(impl StackMatches<N> + StackProgram<N>),
-        registers: &mut impl RegistersLoad<N>,
+        registers: &mut (impl RegistersLoad<N> + RegistersStore<N>),
     ) -> Result<()> {
         // Ensure the number of operands is correct.
         if self.operands.len() != 2 {
@@ -79,21 +91,14 @@ impl<N: Network, const VARIANT: u8> AssertInstruction<N, VARIANT> {
         let input_a = registers.load(stack, &self.operands[0])?;
         let input_b = registers.load(stack, &self.operands[1])?;
 
-        // Assert the inputs.
-        match VARIANT {
-            0 => {
-                if input_a != input_b {
-                    bail!("'{}' failed: '{input_a}' is not equal to '{input_b}' (should be equal)", Self::opcode())
-                }
-            }
-            1 => {
-                if input_a == input_b {
-                    bail!("'{}' failed: '{input_a}' is equal to '{input_b}' (should not be equal)", Self::opcode())
-                }
-            }
-            _ => bail!("Invalid 'assert' variant: {VARIANT}"),
-        }
-        Ok(())
+        // Check the inputs.
+        let output = match VARIANT {
+            0 => Literal::Boolean(Boolean::new(input_a == input_b)),
+            1 => Literal::Boolean(Boolean::new(input_a != input_b)),
+            _ => bail!("Invalid 'is' variant: {VARIANT}"),
+        };
+        // Store the output.
+        registers.store(stack, &self.destination, Value::Plaintext(Plaintext::from(output)))
     }
 
     /// Executes the instruction.
@@ -101,7 +106,7 @@ impl<N: Network, const VARIANT: u8> AssertInstruction<N, VARIANT> {
     pub fn execute<A: circuit::Aleo<Network = N>>(
         &self,
         stack: &(impl StackMatches<N> + StackProgram<N>),
-        registers: &mut impl RegistersLoadCircuit<N, A>,
+        registers: &mut (impl RegistersLoadCircuit<N, A> + RegistersStoreCircuit<N, A>),
     ) -> Result<()> {
         // Ensure the number of operands is correct.
         if self.operands.len() != 2 {
@@ -112,13 +117,16 @@ impl<N: Network, const VARIANT: u8> AssertInstruction<N, VARIANT> {
         let input_a = registers.load_circuit(stack, &self.operands[0])?;
         let input_b = registers.load_circuit(stack, &self.operands[1])?;
 
-        // Assert the inputs.
-        match VARIANT {
-            0 => A::assert(input_a.is_equal(&input_b)),
-            1 => A::assert(input_a.is_not_equal(&input_b)),
-            _ => bail!("Invalid 'assert' variant: {VARIANT}"),
-        }
-        Ok(())
+        // Check the inputs.
+        let output = match VARIANT {
+            0 => circuit::Literal::Boolean(input_a.is_equal(&input_b)),
+            1 => circuit::Literal::Boolean(input_a.is_not_equal(&input_b)),
+            _ => bail!("Invalid 'is' variant: {VARIANT}"),
+        };
+        // Convert the output to a stack value.
+        let output = circuit::Value::Plaintext(circuit::Plaintext::Literal(output, Default::default()));
+        // Store the output.
+        registers.store_circuit(stack, &self.destination, output)
     }
 
     /// Finalizes the instruction.
@@ -126,7 +134,7 @@ impl<N: Network, const VARIANT: u8> AssertInstruction<N, VARIANT> {
     pub fn finalize(
         &self,
         stack: &(impl StackMatches<N> + StackProgram<N>),
-        registers: &mut impl RegistersLoad<N>,
+        registers: &mut (impl RegistersLoad<N> + RegistersStore<N>),
     ) -> Result<()> {
         self.evaluate(stack, registers)
     }
@@ -157,13 +165,13 @@ impl<N: Network, const VARIANT: u8> AssertInstruction<N, VARIANT> {
         }
 
         match VARIANT {
-            0 | 1 => Ok(vec![]),
-            _ => bail!("Invalid 'assert' variant: {VARIANT}"),
+            0 | 1 => Ok(vec![RegisterType::Plaintext(PlaintextType::Literal(LiteralType::Boolean))]),
+            _ => bail!("Invalid 'is' variant: {VARIANT}"),
         }
     }
 }
 
-impl<N: Network, const VARIANT: u8> Parser for AssertInstruction<N, VARIANT> {
+impl<N: Network, const VARIANT: u8> Parser for IsInstruction<N, VARIANT> {
     /// Parses a string into an operation.
     #[inline]
     fn parse(string: &str) -> ParserResult<Self> {
@@ -177,12 +185,20 @@ impl<N: Network, const VARIANT: u8> Parser for AssertInstruction<N, VARIANT> {
         let (string, _) = Sanitizer::parse_whitespaces(string)?;
         // Parse the second operand from the string.
         let (string, second) = Operand::parse(string)?;
+        // Parse the whitespace from the string.
+        let (string, _) = Sanitizer::parse_whitespaces(string)?;
+        // Parse the "into" from the string.
+        let (string, _) = tag("into")(string)?;
+        // Parse the whitespace from the string.
+        let (string, _) = Sanitizer::parse_whitespaces(string)?;
+        // Parse the destination register from the string.
+        let (string, destination) = Register::parse(string)?;
 
-        Ok((string, Self { operands: vec![first, second] }))
+        Ok((string, Self { operands: vec![first, second], destination }))
     }
 }
 
-impl<N: Network, const VARIANT: u8> FromStr for AssertInstruction<N, VARIANT> {
+impl<N: Network, const VARIANT: u8> FromStr for IsInstruction<N, VARIANT> {
     type Err = Error;
 
     /// Parses a string into an operation.
@@ -200,14 +216,14 @@ impl<N: Network, const VARIANT: u8> FromStr for AssertInstruction<N, VARIANT> {
     }
 }
 
-impl<N: Network, const VARIANT: u8> Debug for AssertInstruction<N, VARIANT> {
+impl<N: Network, const VARIANT: u8> Debug for IsInstruction<N, VARIANT> {
     /// Prints the operation as a string.
     fn fmt(&self, f: &mut Formatter) -> fmt::Result {
         Display::fmt(self, f)
     }
 }
 
-impl<N: Network, const VARIANT: u8> Display for AssertInstruction<N, VARIANT> {
+impl<N: Network, const VARIANT: u8> Display for IsInstruction<N, VARIANT> {
     /// Prints the operation to a string.
     fn fmt(&self, f: &mut Formatter) -> fmt::Result {
         // Ensure the number of operands is 2.
@@ -217,11 +233,12 @@ impl<N: Network, const VARIANT: u8> Display for AssertInstruction<N, VARIANT> {
         }
         // Print the operation.
         write!(f, "{} ", Self::opcode())?;
-        self.operands.iter().try_for_each(|operand| write!(f, "{operand} "))
+        self.operands.iter().try_for_each(|operand| write!(f, "{operand} "))?;
+        write!(f, "into {}", self.destination)
     }
 }
 
-impl<N: Network, const VARIANT: u8> FromBytes for AssertInstruction<N, VARIANT> {
+impl<N: Network, const VARIANT: u8> FromBytes for IsInstruction<N, VARIANT> {
     /// Reads the operation from a buffer.
     fn read_le<R: Read>(mut reader: R) -> IoResult<Self> {
         // Initialize the vector for the operands.
@@ -230,13 +247,15 @@ impl<N: Network, const VARIANT: u8> FromBytes for AssertInstruction<N, VARIANT> 
         for _ in 0..2 {
             operands.push(Operand::read_le(&mut reader)?);
         }
+        // Read the destination register.
+        let destination = Register::read_le(&mut reader)?;
 
         // Return the operation.
-        Ok(Self { operands })
+        Ok(Self { operands, destination })
     }
 }
 
-impl<N: Network, const VARIANT: u8> ToBytes for AssertInstruction<N, VARIANT> {
+impl<N: Network, const VARIANT: u8> ToBytes for IsInstruction<N, VARIANT> {
     /// Writes the operation to a buffer.
     fn write_le<W: Write>(&self, mut writer: W) -> IoResult<()> {
         // Ensure the number of operands is 2.
@@ -244,22 +263,21 @@ impl<N: Network, const VARIANT: u8> ToBytes for AssertInstruction<N, VARIANT> {
             return Err(error(format!("The number of operands must be 2, found {}", self.operands.len())));
         }
         // Write the operands.
-        self.operands.iter().try_for_each(|operand| operand.write_le(&mut writer))
+        self.operands.iter().try_for_each(|operand| operand.write_le(&mut writer))?;
+        // Write the destination register.
+        self.destination.write_le(&mut writer)
     }
 }
 
 #[cfg(test)]
 mod tests {
     use super::*;
-    use crate::{
-        process::Stack,
-        program::test_helpers::{sample_finalize_registers, sample_registers},
+    use crate::process::{
+        instruction::test_helpers::{sample_finalize_registers, sample_registers},
+        Stack,
     };
     use circuit::AleoV0;
-    use console::{
-        network::Testnet3,
-        program::{Identifier, Literal, LiteralType},
-    };
+    use console::{network::Testnet3, program::Identifier};
     use snarkvm_synthesizer_snark::{ProvingKey, VerifyingKey};
 
     use std::collections::HashMap;
@@ -270,6 +288,7 @@ mod tests {
     const ITERATIONS: usize = 100;
 
     /// Samples the stack. Note: Do not replicate this for real program use, it is insecure.
+    #[allow(clippy::type_complexity)]
     fn sample_stack(
         opcode: Opcode,
         type_a: LiteralType,
@@ -277,7 +296,7 @@ mod tests {
         mode_a: circuit::Mode,
         mode_b: circuit::Mode,
         cache: &mut HashMap<String, (ProvingKey<CurrentNetwork>, VerifyingKey<CurrentNetwork>)>,
-    ) -> Result<(Stack<CurrentNetwork>, Vec<Operand<CurrentNetwork>>)> {
+    ) -> Result<(Stack<CurrentNetwork>, Vec<Operand<CurrentNetwork>>, Register<CurrentNetwork>)> {
         use crate::{Process, Program};
 
         // Initialize the opcode.
@@ -289,6 +308,7 @@ mod tests {
         // Initialize the registers.
         let r0 = Register::Locator(0);
         let r1 = Register::Locator(1);
+        let r2 = Register::Locator(2);
 
         // Initialize the program.
         let program = Program::from_str(&format!(
@@ -296,13 +316,13 @@ mod tests {
             function {function_name}:
                 input {r0} as {type_a}.{mode_a};
                 input {r1} as {type_b}.{mode_b};
-                {opcode} {r0} {r1};
+                {opcode} {r0} {r1} into {r2};
                 finalize {r0} {r1};
 
             finalize {function_name}:
                 input {r0} as {type_a}.public;
                 input {r1} as {type_b}.public;
-                {opcode} {r0} {r1};
+                {opcode} {r0} {r1} into {r2};
         "
         ))?;
 
@@ -314,11 +334,14 @@ mod tests {
         // Initialize the stack.
         let stack = Stack::new(&Process::load_with_cache(cache)?, &program)?;
 
-        Ok((stack, operands))
+        Ok((stack, operands, r2))
     }
 
-    fn check_assert<const VARIANT: u8>(
-        operation: impl FnOnce(Vec<Operand<CurrentNetwork>>) -> AssertInstruction<CurrentNetwork, VARIANT>,
+    fn check_is<const VARIANT: u8>(
+        operation: impl FnOnce(
+            Vec<Operand<CurrentNetwork>>,
+            Register<CurrentNetwork>,
+        ) -> IsInstruction<CurrentNetwork, VARIANT>,
         opcode: Opcode,
         literal_a: &Literal<CurrentNetwork>,
         literal_b: &Literal<CurrentNetwork>,
@@ -326,6 +349,8 @@ mod tests {
         mode_b: &circuit::Mode,
         cache: &mut HashMap<String, (ProvingKey<CurrentNetwork>, VerifyingKey<CurrentNetwork>)>,
     ) {
+        use circuit::Eject;
+
         println!("Checking '{opcode}' for '{literal_a}.{mode_a}' and '{literal_b}.{mode_b}'");
 
         // Initialize the types.
@@ -334,58 +359,74 @@ mod tests {
         assert_eq!(type_a, type_b, "The two literals must be the *same* type for this test");
 
         // Initialize the stack.
-        let (stack, operands) = sample_stack(opcode, type_a, type_b, *mode_a, *mode_b, cache).unwrap();
+        let (stack, operands, destination) = sample_stack(opcode, type_a, type_b, *mode_a, *mode_b, cache).unwrap();
         // Initialize the operation.
-        let operation = operation(operands);
+        let operation = operation(operands, destination.clone());
         // Initialize the function name.
         let function_name = Identifier::from_str("run").unwrap();
+        // Initialize a destination operand.
+        let destination_operand = Operand::Register(destination);
 
         /* First, check the operation *succeeds* when both operands are `literal_a.mode_a`. */
         {
             // Attempt to compute the valid operand case.
             let values = [(literal_a, None), (literal_a, None)];
             let mut registers = sample_registers(&stack, &function_name, &values).unwrap();
-            let result_a = operation.evaluate(&stack, &mut registers);
+            operation.evaluate(&stack, &mut registers).unwrap();
 
-            // Ensure the result is correct.
-            match VARIANT {
-                0 => assert!(result_a.is_ok(), "Instruction '{operation}' failed (console): {literal_a} {literal_a}"),
-                1 => assert!(
-                    result_a.is_err(),
-                    "Instruction '{operation}' should have failed (console): {literal_a} {literal_a}"
-                ),
-                _ => panic!("Found an invalid 'assert' variant in the test"),
+            // Retrieve the output.
+            let output_a = registers.load_literal(&stack, &destination_operand).unwrap();
+
+            // Ensure the output is correct.
+            if let Literal::Boolean(output_a) = output_a {
+                match VARIANT {
+                    0 => assert!(*output_a, "Instruction '{operation}' failed (console): {literal_a} {literal_a}"),
+                    1 => assert!(
+                        !*output_a,
+                        "Instruction '{operation}' should have failed (console): {literal_a} {literal_a}"
+                    ),
+                    _ => panic!("Found an invalid 'is' variant in the test"),
+                }
+            } else {
+                panic!("The output must be a boolean (console)");
             }
 
             // Attempt to compute the valid operand case.
             let values = [(literal_a, Some(*mode_a)), (literal_a, Some(*mode_a))];
             let mut registers = sample_registers(&stack, &function_name, &values).unwrap();
-            let result_b = operation.execute::<CurrentAleo>(&stack, &mut registers);
+            operation.execute::<CurrentAleo>(&stack, &mut registers).unwrap();
 
-            // Ensure the result is correct.
-            match VARIANT {
-                0 => assert!(
-                    result_b.is_ok(),
-                    "Instruction '{operation}' failed (circuit): {literal_a}.{mode_a} {literal_a}.{mode_a}"
-                ),
-                1 => assert!(
-                    result_b.is_ok(), // <- 'is_ok()' is correct! The circuit should execute, but the constraint should not be satisfied.
-                    "Instruction '{operation}' should not have panicked (circuit): {literal_a}.{mode_a} {literal_a}.{mode_a}"
-                ),
-                _ => panic!("Found an invalid 'assert' variant in the test"),
+            // Retrieve the output.
+            let output_b = registers.load_literal_circuit(&stack, &destination_operand).unwrap();
+
+            // Ensure the output is correct.
+            if let circuit::Literal::Boolean(output_b) = output_b {
+                match VARIANT {
+                    0 => assert!(
+                        output_b.eject_value(),
+                        "Instruction '{operation}' failed (circuit): {literal_a}.{mode_a} {literal_a}.{mode_a}"
+                    ),
+                    1 => assert!(
+                        !output_b.eject_value(),
+                        "Instruction '{operation}' should have failed (circuit): {literal_a}.{mode_a} {literal_a}.{mode_a}"
+                    ),
+                    _ => panic!("Found an invalid 'is' variant in the test"),
+                }
+            } else {
+                panic!("The output must be a boolean (circuit)");
             }
 
-            // Ensure the circuit is correct.
+            // Ensure the circuit is satisfied.
             match VARIANT {
                 0 => assert!(
                     <CurrentAleo as circuit::Environment>::is_satisfied(),
                     "Instruction '{operation}' should be satisfied (circuit): {literal_a}.{mode_a} {literal_a}.{mode_a}"
                 ),
                 1 => assert!(
-                    !<CurrentAleo as circuit::Environment>::is_satisfied(),
-                    "Instruction '{operation}' should not be satisfied (circuit): {literal_a}.{mode_a} {literal_a}.{mode_a}"
+                    <CurrentAleo as circuit::Environment>::is_satisfied(),
+                    "Instruction '{operation}' should be satisfied (circuit): {literal_a}.{mode_a} {literal_a}.{mode_a}"
                 ),
-                _ => panic!("Found an invalid 'assert' variant in the test"),
+                _ => panic!("Found an invalid 'is' variant in the test"),
             }
 
             // Reset the circuit.
@@ -393,16 +434,23 @@ mod tests {
 
             // Attempt to finalize the valid operand case.
             let mut registers = sample_finalize_registers(&stack, &function_name, &[literal_a, literal_a]).unwrap();
-            let result_c = operation.finalize(&stack, &mut registers);
+            operation.finalize(&stack, &mut registers).unwrap();
 
-            // Ensure the result is correct.
-            match VARIANT {
-                0 => assert!(result_c.is_ok(), "Instruction '{operation}' failed (console): {literal_a} {literal_a}"),
-                1 => assert!(
-                    result_c.is_err(),
-                    "Instruction '{operation}' should have failed (console): {literal_a} {literal_a}"
-                ),
-                _ => panic!("Found an invalid 'assert' variant in the test"),
+            // Retrieve the output.
+            let output_c = registers.load_literal(&stack, &destination_operand).unwrap();
+
+            // Ensure the output is correct.
+            if let Literal::Boolean(output_c) = output_c {
+                match VARIANT {
+                    0 => assert!(*output_c, "Instruction '{operation}' failed (finalize): {literal_a} {literal_a}"),
+                    1 => assert!(
+                        !*output_c,
+                        "Instruction '{operation}' should have failed (finalize): {literal_a} {literal_a}"
+                    ),
+                    _ => panic!("Found an invalid 'is' variant in the test"),
+                }
+            } else {
+                panic!("The output must be a boolean (finalize)");
             }
         }
         /* Next, check the mismatching literals *fail*. */
@@ -410,47 +458,61 @@ mod tests {
             // Attempt to compute the valid operand case.
             let values = [(literal_a, None), (literal_b, None)];
             let mut registers = sample_registers(&stack, &function_name, &values).unwrap();
-            let result_a = operation.evaluate(&stack, &mut registers);
+            operation.evaluate(&stack, &mut registers).unwrap();
 
-            // Ensure the result is correct.
-            match VARIANT {
-                0 => assert!(
-                    result_a.is_err(),
-                    "Instruction '{operation}' should have failed (console): {literal_a} {literal_b}"
-                ),
-                1 => assert!(result_a.is_ok(), "Instruction '{operation}' failed (console): {literal_a} {literal_b}"),
-                _ => panic!("Found an invalid 'assert' variant in the test"),
+            // Retrieve the output.
+            let output_a = registers.load_literal(&stack, &destination_operand).unwrap();
+
+            // Ensure the output is correct.
+            if let Literal::Boolean(output_a) = output_a {
+                match VARIANT {
+                    0 => assert!(
+                        !*output_a,
+                        "Instruction '{operation}' should have failed (console): {literal_a} {literal_b}"
+                    ),
+                    1 => assert!(*output_a, "Instruction '{operation}' failed (console): {literal_a} {literal_b}"),
+                    _ => panic!("Found an invalid 'is' variant in the test"),
+                }
+            } else {
+                panic!("The output must be a boolean (console)");
             }
 
             // Attempt to compute the valid operand case.
             let values = [(literal_a, Some(*mode_a)), (literal_b, Some(*mode_b))];
             let mut registers = sample_registers(&stack, &function_name, &values).unwrap();
-            let result_b = operation.execute::<CurrentAleo>(&stack, &mut registers);
+            operation.execute::<CurrentAleo>(&stack, &mut registers).unwrap();
 
-            // Ensure the result is correct.
-            match VARIANT {
-                0 => assert!(
-                    result_b.is_ok(), // <- 'is_ok()' is correct! The circuit should execute, but the constraint should not be satisfied.
-                    "Instruction '{operation}' should not have panicked (circuit): {literal_a}.{mode_a} {literal_b}.{mode_b}"
-                ),
-                1 => assert!(
-                    result_b.is_ok(),
-                    "Instruction '{operation}' failed (circuit): {literal_a}.{mode_a} {literal_b}.{mode_b}"
-                ),
-                _ => panic!("Found an invalid 'assert' variant in the test"),
+            // Retrieve the output.
+            let output_b = registers.load_literal_circuit(&stack, &destination_operand).unwrap();
+
+            // Ensure the output is correct.
+            if let circuit::Literal::Boolean(output_b) = output_b {
+                match VARIANT {
+                    0 => assert!(
+                        !output_b.eject_value(),
+                        "Instruction '{operation}' failed (circuit): {literal_a}.{mode_a} {literal_b}.{mode_b}"
+                    ),
+                    1 => assert!(
+                        output_b.eject_value(),
+                        "Instruction '{operation}' should have failed (circuit): {literal_a}.{mode_a} {literal_b}.{mode_b}"
+                    ),
+                    _ => panic!("Found an invalid 'is' variant in the test"),
+                }
+            } else {
+                panic!("The output must be a boolean (circuit)");
             }
 
             // Ensure the circuit is correct.
             match VARIANT {
                 0 => assert!(
-                    !<CurrentAleo as circuit::Environment>::is_satisfied(),
-                    "Instruction '{operation}' should not be satisfied (circuit): {literal_a}.{mode_a} {literal_b}.{mode_b}"
+                    <CurrentAleo as circuit::Environment>::is_satisfied(),
+                    "Instruction '{operation}' should be satisfied (circuit): {literal_a}.{mode_a} {literal_b}.{mode_b}"
                 ),
                 1 => assert!(
                     <CurrentAleo as circuit::Environment>::is_satisfied(),
                     "Instruction '{operation}' should be satisfied (circuit): {literal_a}.{mode_a} {literal_b}.{mode_b}"
                 ),
-                _ => panic!("Found an invalid 'assert' variant in the test"),
+                _ => panic!("Found an invalid 'is' variant in the test"),
             }
 
             // Reset the circuit.
@@ -458,21 +520,28 @@ mod tests {
 
             // Attempt to finalize the valid operand case.
             let mut registers = sample_finalize_registers(&stack, &function_name, &[literal_a, literal_b]).unwrap();
-            let result_c = operation.finalize(&stack, &mut registers);
+            operation.finalize(&stack, &mut registers).unwrap();
 
-            // Ensure the result is correct.
-            match VARIANT {
-                0 => assert!(
-                    result_c.is_err(),
-                    "Instruction '{operation}' should have failed (console): {literal_a} {literal_b}"
-                ),
-                1 => assert!(result_c.is_ok(), "Instruction '{operation}' failed (console): {literal_a} {literal_b}"),
-                _ => panic!("Found an invalid 'assert' variant in the test"),
+            // Retrieve the output.
+            let output_c = registers.load_literal(&stack, &destination_operand).unwrap();
+
+            // Ensure the output is correct.
+            if let Literal::Boolean(output_c) = output_c {
+                match VARIANT {
+                    0 => assert!(
+                        !*output_c,
+                        "Instruction '{operation}' should have failed (finalize): {literal_a} {literal_b}"
+                    ),
+                    1 => assert!(*output_c, "Instruction '{operation}' failed (finalize): {literal_a} {literal_b}"),
+                    _ => panic!("Found an invalid 'is' variant in the test"),
+                }
+            } else {
+                panic!("The output must be a boolean (finalize)");
             }
         }
     }
 
-    fn check_assert_fails(
+    fn check_is_fails(
         opcode: Opcode,
         literal_a: &Literal<CurrentNetwork>,
         literal_b: &Literal<CurrentNetwork>,
@@ -494,11 +563,11 @@ mod tests {
     }
 
     #[test]
-    fn test_assert_eq_succeeds() {
+    fn test_is_eq_succeeds() {
         // Initialize the operation.
-        let operation = |operands| AssertEq::<CurrentNetwork> { operands };
+        let operation = |operands, destination| IsEq::<CurrentNetwork> { operands, destination };
         // Initialize the opcode.
-        let opcode = AssertEq::<CurrentNetwork>::opcode();
+        let opcode = IsEq::<CurrentNetwork>::opcode();
 
         // Prepare the rng.
         let mut rng = TestRng::default();
@@ -517,7 +586,7 @@ mod tests {
                 for mode_a in &modes_a {
                     for mode_b in &modes_b {
                         // Check the operation.
-                        check_assert(operation, opcode, literal_a, literal_b, mode_a, mode_b, &mut cache);
+                        check_is(operation, opcode, literal_a, literal_b, mode_a, mode_b, &mut cache);
                     }
                 }
             }
@@ -525,9 +594,9 @@ mod tests {
     }
 
     #[test]
-    fn test_assert_eq_fails() {
+    fn test_is_eq_fails() {
         // Initialize the opcode.
-        let opcode = AssertEq::<CurrentNetwork>::opcode();
+        let opcode = IsEq::<CurrentNetwork>::opcode();
 
         // Prepare the rng.
         let mut rng = TestRng::default();
@@ -548,7 +617,7 @@ mod tests {
                         for mode_a in &modes_a {
                             for mode_b in &modes_b {
                                 // Check the operation fails.
-                                check_assert_fails(opcode, literal_a, literal_b, mode_a, mode_b, &mut cache);
+                                check_is_fails(opcode, literal_a, literal_b, mode_a, mode_b, &mut cache);
                             }
                         }
                     }
@@ -558,11 +627,11 @@ mod tests {
     }
 
     #[test]
-    fn test_assert_neq_succeeds() {
+    fn test_is_neq_succeeds() {
         // Initialize the operation.
-        let operation = |operands| AssertNeq::<CurrentNetwork> { operands };
+        let operation = |operands, destination| IsNeq::<CurrentNetwork> { operands, destination };
         // Initialize the opcode.
-        let opcode = AssertNeq::<CurrentNetwork>::opcode();
+        let opcode = IsNeq::<CurrentNetwork>::opcode();
 
         // Prepare the rng.
         let mut rng = TestRng::default();
@@ -581,7 +650,7 @@ mod tests {
                 for mode_a in &modes_a {
                     for mode_b in &modes_b {
                         // Check the operation.
-                        check_assert(operation, opcode, literal_a, literal_b, mode_a, mode_b, &mut cache);
+                        check_is(operation, opcode, literal_a, literal_b, mode_a, mode_b, &mut cache);
                     }
                 }
             }
@@ -589,9 +658,9 @@ mod tests {
     }
 
     #[test]
-    fn test_assert_neq_fails() {
+    fn test_is_neq_fails() {
         // Initialize the opcode.
-        let opcode = AssertNeq::<CurrentNetwork>::opcode();
+        let opcode = IsNeq::<CurrentNetwork>::opcode();
 
         // Prepare the rng.
         let mut rng = TestRng::default();
@@ -612,7 +681,7 @@ mod tests {
                         for mode_a in &modes_a {
                             for mode_b in &modes_b {
                                 // Check the operation fails.
-                                check_assert_fails(opcode, literal_a, literal_b, mode_a, mode_b, &mut cache);
+                                check_is_fails(opcode, literal_a, literal_b, mode_a, mode_b, &mut cache);
                             }
                         }
                     }
@@ -623,16 +692,18 @@ mod tests {
 
     #[test]
     fn test_parse() {
-        let (string, assert) = AssertEq::<CurrentNetwork>::parse("assert.eq r0 r1").unwrap();
+        let (string, is) = IsEq::<CurrentNetwork>::parse("is.eq r0 r1 into r2").unwrap();
         assert!(string.is_empty(), "Parser did not consume all of the string: '{string}'");
-        assert_eq!(assert.operands.len(), 2, "The number of operands is incorrect");
-        assert_eq!(assert.operands[0], Operand::Register(Register::Locator(0)), "The first operand is incorrect");
-        assert_eq!(assert.operands[1], Operand::Register(Register::Locator(1)), "The second operand is incorrect");
+        assert_eq!(is.operands.len(), 2, "The number of operands is incorrect");
+        assert_eq!(is.operands[0], Operand::Register(Register::Locator(0)), "The first operand is incorrect");
+        assert_eq!(is.operands[1], Operand::Register(Register::Locator(1)), "The second operand is incorrect");
+        assert_eq!(is.destination, Register::Locator(2), "The destination register is incorrect");
 
-        let (string, assert) = AssertNeq::<CurrentNetwork>::parse("assert.neq r0 r1").unwrap();
+        let (string, is) = IsNeq::<CurrentNetwork>::parse("is.neq r0 r1 into r2").unwrap();
         assert!(string.is_empty(), "Parser did not consume all of the string: '{string}'");
-        assert_eq!(assert.operands.len(), 2, "The number of operands is incorrect");
-        assert_eq!(assert.operands[0], Operand::Register(Register::Locator(0)), "The first operand is incorrect");
-        assert_eq!(assert.operands[1], Operand::Register(Register::Locator(1)), "The second operand is incorrect");
+        assert_eq!(is.operands.len(), 2, "The number of operands is incorrect");
+        assert_eq!(is.operands[0], Operand::Register(Register::Locator(0)), "The first operand is incorrect");
+        assert_eq!(is.operands[1], Operand::Register(Register::Locator(1)), "The second operand is incorrect");
+        assert_eq!(is.destination, Register::Locator(2), "The destination register is incorrect");
     }
 }
