+++ synthesizer/src/process/instruction/operation/macros.rs
