+++ synthesizer/src/process/instruction/operation/mod.rs
