+++ synthesizer/src/program/closure/parse.rs
@@ -14,7 +14,7 @@
 
 use super::*;
 
-impl<N: Network> Parser for Closure<N> {
+impl<N: Network, Instruction: InstructionTrait<N>> Parser for ClosureCore<N, Instruction> {
     /// Parses a string into a closure.
     #[inline]
     fn parse(string: &str) -> ParserResult<Self> {
@@ -49,7 +49,7 @@ impl<N: Network> Parser for Closure<N> {
     }
 }
 
-impl<N: Network> FromStr for Closure<N> {
+impl<N: Network, Instruction: InstructionTrait<N>> FromStr for ClosureCore<N, Instruction> {
     type Err = Error;
 
     /// Returns a closure from a string literal.
@@ -66,14 +66,14 @@ impl<N: Network> FromStr for Closure<N> {
     }
 }
 
-impl<N: Network> Debug for Closure<N> {
+impl<N: Network, Instruction: InstructionTrait<N>> Debug for ClosureCore<N, Instruction> {
     /// Prints the closure as a string.
     fn fmt(&self, f: &mut Formatter) -> fmt::Result {
         Display::fmt(self, f)
     }
 }
 
-impl<N: Network> Display for Closure<N> {
+impl<N: Network, Instruction: InstructionTrait<N>> Display for ClosureCore<N, Instruction> {
     /// Prints the closure as a string.
     fn fmt(&self, f: &mut Formatter) -> fmt::Result {
         // Write the closure to a string.
@@ -87,6 +87,7 @@ impl<N: Network> Display for Closure<N> {
 #[cfg(test)]
 mod tests {
     use super::*;
+    use crate::process::Closure;
     use console::network::Testnet3;
 
     type CurrentNetwork = Testnet3;
